Refactor EMR
